Move fibonacciPositions into geometry module
