+++ plato/draw/fresnel/Ellipsoids.py
@@ -2,16 +2,9 @@ import fresnel
 import itertools
 import numpy as np
 from ... import draw
+from ...geometry import fibonacciPositions
 from ..internal import ShapeAttribute, ShapeDecorator
 from .FresnelPrimitive import FresnelPrimitive
-
-def fibonacciPositions(n_b, a=.5, b=0.5, c=0.5):
-    m = np.arange(n_b).astype(np.float32)
-    phi = m*np.pi*(3 - np.sqrt(5))
-    vy = 2*m/n_b + 1/n_b - 1
-    return np.array([a*np.sqrt(1 - vy**2)*np.cos(phi),
-                     b*vy, c*np.sqrt(1 - vy**2)*np.sin(phi)]).T
-
 
 @ShapeDecorator
 class Ellipsoids(draw.Ellipsoids, FresnelPrimitive):
